Extract ignore, unignore to ignore
